requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
groq>=0.13.0
python-dotenv>=1.0.1
pinecone>=6.0.0
huggingface-hub>=0.20.0
